refactor: Simplify package structure by moving BaseEntity
