+++ backend/demo/src/main/java/com/sesac/solbid/domain/ProductImage.java
@@ -1,6 +1,5 @@
 package com.sesac.solbid.domain;
 
-import com.sesac.solbid.domain.baseentity.BaseEntity;
 import jakarta.persistence.*;
 import lombok.*;
